+++ examples/libredaq_example_pga_adc.cpp
@@ -25,7 +25,7 @@ void my_callback_ADC(const libredaq::TCallbackData_ADC& data)
 #if 1
     static FILE* f = fopen("adc.txt", "wt");
     fprintf(f, "%15.7f", data.device_timestamp);
-    for (int k = 0; k < data.adc_data_volts.size(); k++)
+    for (size_t k = 0; k < data.adc_data_volts.size(); k++)
         fprintf(f, " %5.05f", data.adc_data_volts[k]);
     fprintf(f, "\n");
 #endif
